Reorganize repository, add docs, run pylint
