+++ .env.example
@@ -8,12 +8,12 @@ DB_HOST=localhost
 DB_NAME=azure
 DB_PORT=5999
 DB_USERNAME=root
-DB_PASSWORD=your_password
+DB_PASSWORD=localpassword
 
 # For local development when we spin up a fresh postgres container
 POSTGRES_DB=azure
 POSTGRES_USER=root
-POSTGRES_PASSWORD=your_password
+POSTGRES_PASSWORD=localpassword
 
 # Azure API
 APP_ID=J8N8I8KE2Y
