+++ archive/hummingbird/test_data/formatted_multi_variant_product.json
