+++ archive/hummingbird/test_data/formatted_single_variant_product.json
