+++ src/settings.py
@@ -1,3 +1,5 @@
+"""Application settings imported from ENV files"""
+
 import json
 import os
 from pathlib import Path
@@ -6,13 +8,14 @@ from pydantic import field_validator
 from pydantic_settings import BaseSettings, SettingsConfigDict
 
 # Determine which env files to load based on ENV variable
-BASE_DIR = Path(__file__).resolve().parent
+BASE_DIR = Path(__file__).resolve().parent.parent
+
 env = os.getenv("ENV")
 env_files = []
+
 if env:
     env_files.append(str(BASE_DIR / f".env.{env}"))
 env_files.extend([str(BASE_DIR / ".env.local"), str(BASE_DIR / ".env")])
-
 
 class Settings(BaseSettings):
     """Application settings loaded from environment files."""
